[analysis/dense/gpcca/_indexsearch] index now np.int64

+++ msmtools/analysis/dense/gpcca.py
@@ -376,7 +376,7 @@ def _indexsearch(X):
     # local copy of the eigenvectors
     ortho_sys = np.copy(X)
 
-    index = np.zeros(m, dtype=np.int32)
+    index = np.zeros(m, dtype=np.int64)
     max_dist = 0.0                     
     
     # First vertex: row with largest norm.
